Initial commit: 6502 CPU work in progress implementation and debugging
So far, I have implemented all of the 6502 processor functionality, but
it still needs to be debugged and fixed, since the NESTest ROM is not
producing the right behavior. I've implemented basic CPU logging so that
I can start comparing it to the NESTest golden log, and find where it
goes wrong. (Probably in the stack related instructions - JSR, RTS...)
Before all that, I need to figure out why address 0x0 in RAM is always
0xFF at the start. (Likely a bug in the Bus class)

I'm working on the project using Visual Studio on Windows 10 currently,
so that is the environment that's set up at the moment, but setting up
a better build environment, one that's cross-platform, is on my to-do
for very soon.