Logging, Build: Replaced all logging with fmtlib, Improved project structure
Successfully added fmtlib to the project and replaced all the prints, and
file I/O with it. This makes the code cross-platform now, since it's not
using any "safe" versions of stdio functions like printf_s(), but it's still
using safe and fast functions for doing logging.

Any new libraries I include to the project should go into the lib/ folder
under the appropriate name for the library. Also made it so that Visual
Studio now outputs to the bin/Windows/ directory, with sub folders for
different build targets (x64, x86) and moved the Debug folder in there too.

Next I'll work on making a bare bones build system for Linux, so that I can
work on the project and compile it on there.