Core: Emulator structure, Started PPU
I've moved most of the functionality out of the `main()` funciton
and into the NesCore class. Started drafting the PPU classes and
added it and its bus to the NesCore. Should now continue to work
on the project structure and interaction between all the classes.
That probably will include making an IBus interface and trying out
a different concrete implementation of the bus class, in order to
be more efficient and support some functionality that I need which
is currently missing from the Bus class.