Core, Bus: Bit of restructuring, new bus class
Continued restructuring the project. Made an IBus interface and
started using it everywhere instead of the Bus class. The previous
Bus class was renamed to NesMultiMapBus, and we currently use that
as the concrete implementation for the bus. I started working on
a new implementation `NesVectorBus` which might be better to use
overall, however currently it doesn't work as intended and needs
to be debugged. TODO: Fix `NesVectorBus` class to work propperly.