Cartridge: We now load data into the Cartridge
Migrated the whole system to use the `NesCartridge` class to load
ROMs and interact with the rest of the system. Added some methods
to `NesCore` for further future implementation, and changed
`nesTest()` to use the Cartridge.

Bus no longer translates addresses, instead just passes them to
appropriate slaves as-is and lets them handle the translation.